Add a project after correction 1) Перенести PageObject в src/main 2) Разделить тесты заказа и вопросов на два разных тестовых класса
